wifi: add initial OpenDroneID WiFi sender reference code
The reference code allows to send WiFi NAN frames according to the
current position acquired by a GPS receiver. The GPS receiver is queried
using the open source gpsd program. Next to NAN frames, there is also a
function implemented to change SSIDs periodically (not following the
OpenDroneID spec, but displaying some information in a colon seperated
format).

In tests performed by Kaleb Kleine and Anratech, it this reference
implementation is not yet sufficient for Android Phones (e.g. Pixel 2).
Those phones seem to expect special NAN beacons to be sent as well,
which is not part of this reference implementation at the moment.

Signed-off-by: Marek Sobe <ms@simonwunderlich.de>
Signed-off-by: Simon Wunderlich <sw@simonwunderlich.de>