Remove save_copyfill functionality
Remove the save_copyfill function since it doesn't add any
additional protection. It will blindly copy whatever size
of bytes that the caller asks it to, without having any
possibility to check whether it fits in the destination
buffer or not.

The caller must instead look at the data buffers being
filled and ensure it only fills in the amount of data that
fits.

If we want such functionality, there needs to be a dedicated
copyfill function for each message/buffer-type instead.

Make the rest of the helper functions static, since they
are not meant to be used outside the core-c implementation.

Also reduce the ID data in test_inout.c by one byte since
it was too long to fit.

Signed-off-by: Soren Friis <soren.friis@intel.com>