Add support for Classification data
the ASD-STAN Direct Remote ID standard introduces new fields
for transmitting classification data.

Changed the name of the operator location type to be more
descriptive and at the same time fixed a bug where this
field was encoded with too few bits and in the wrong position.

Also a new status enum value for Emergency has been defined.

A future version of the ASTM Remote ID standard will include
the same new fields and values.

Signed-off-by: Soren Friis <soren.friis@intel.com>