Give a name to struct to allow forward declaration
Functionally it's the same but if you want to make a forward declaration,
you have to give it a name in the tag namespace.
https://stackoverflow.com/questions/612328/difference-between-struct-and-typedef-struct-in-c/612350#612350

Personally, I prefer to omit the typedef and refer to the type directly as
struct, enum or union. The typedef save a little typing, but it hides the
fact that it's a structure type. This would be a big change in the API and
I prefer to align myself with how it was already done.